Feat: Hook up form inputs to generate meme text
Created a function that will handle the change on the form inputs.
This will update the state of the top and bottom text.

Use the state of the top and bottom text to generate text that will
overlay the meme image.

Style the font so that it appears on top of the image.